pyproject.toml
[project]
name = "ai-guide-creator"
version = "1.0.0"
description = "AI-powered guide creation system using CrewAI with Phoenix observability"
authors = [{ name = "AI Guide Creator", email = "contact@example.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.13"
keywords = ["ai", "crewai", "guide-generation", "llm", "observability", "phoenix"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "arize-phoenix>=10.5.0",
    "crewai[tools]>=0.121.1,<1.0.0",
    "openinference-instrumentation-crewai>=0.1.10",
    "openinference-instrumentation-openai>=0.1.0,<=0.1.30",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
guide-creator = "guide_creator_flow.main:kickoff"
kickoff = "guide_creator_flow.main:kickoff"
run_crew = "guide_creator_flow.main:kickoff"
plot = "guide_creator_flow.main:plot"

[project.urls]
Homepage = "https://github.com/tommyyau/ai-guide-creator"
Repository = "https://github.com/tommyyau/ai-guide-creator"
Issues = "https://github.com/tommyyau/ai-guide-creator/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/guide_creator_flow"]

[tool.crewai]
type = "flow"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
